Adds Logic to Find Static Files
I noticed that static files in nested directories--e.g. static/fonts, static/js/vendor--were not being included when this theme was installed as a package. This caused issues with the theme not loading required JavaScript to work property (and also not displaying the right fonts). (One could see this by inspecting the network calls the theme was making and getting lots of 404 errors.)

I've added logic here to automatically add all files in the `static` directory to the theme package. The logic I added does that by recursively traversing the `static` directory and adding all found files' relative paths to the setup import.