fix(ui) LAI2-10606: make the right sidebar menu sticky (#58)
This PR fixes the regression and makes:
- the left menubar scrollable independently of the content
- the right menubar sticky

The regression was introduced when we disabled the footer completely to remove the "black bar" at the bottom of the doc page. The `handleLeftSidebar` code relied on it to compute the height. I removed the `handleLeftSidebar` code completely and set the left sidebar height to `100%`, equaling the result of the computation I removed most of the time. This fixes the crash that was leading to the `handleRightSidebar` function not being called at all. Now that it's being called, the right sidebar is sticky again.